pyproject.toml
[project]
name = "open-tars"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
    "pyautogui>=0.9.54",
]
